Create bubble_sort function
It's the bubble sorth algorithm implementation. Conists in moving the
highest values in the array to the end, one for each iteration.

First it makes a copy to not lose data; then we use two variables to
control the iterations: sorted, that determines when the array is
sorted and max_iterations, that helps do less iterations each time an
element is moved to the end of the arary.

There are two loops, a while and a each_with_index. The first one ends
when the array is sorted and assigns true before it starts iterating the
array; the second one iterates all the array the first time, and then
one element less the next times because it is not necessary to compare
the elements that are already sorted.

In the each_with_index loop, it makes a change when the current element
is greater than the next element, to do that it stores the next
element in a temporary variable, then assign the current to the next
and finally assign the temporary to the current element. Also, when
there is a change, the sorted variable becomes false. Finally, when an
array iteration ends, it substracts one to the max_iterations variable.